refactoring podlister config
