+++ podlister/main.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 	"text/template"
 
+	"podlister/config"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -18,33 +19,9 @@ import (
 	"k8s.io/client-go/rest"
 )
 
-type Config struct {
-	Bucket struct {
-		Key       string `env:"BUCKET_KEY" env-required:"Bucket key"`
-		Secret    string `env:"BUCKET_SECRET" env-required:"Bucket secret"`
-		URL       string `yaml:"url" env:"BUCKET_URL" env-default:"https://nyc3.digitaloceanspaces.com"`
-		Name      string `yaml:"name" env:"BUCKET_NAME" env-required:"Bucket name"`
-		Region    string `yaml:"region" env:"BUCKET_REGION" env-default:"us-east-1"`
-		Privilege string `yaml:"privilege" env:"BUCKET_PRIVILEGE" env-default:"public-read"`
-	} `yaml:"bucket"`
-	Template struct {
-		Name   string `yaml:"name" env:"TEMPLATE_NAME" env-default:"index.template"`
-		Output string `yaml:"output" env:"TEMPLATE_OUTPUT" env-default:"index.html"`
-	} `yaml:"template"`
-	Service struct {
-		Name   string `yaml:"name" env:"SERVICE_NAME" env-required:"Service name"`
-	} `yaml:"service"`
-}
-
-type Endpoint struct {
-	Namespace string
-	Svc       string
-	Ips       []string
-}
-
 var (
 	namespacePath = "/var/run/secrets/kubernetes.io/serviceaccount/namespace"
-	configPath = "./config.yaml"
+	configPath    = "./config/config.yaml"
 )
 
 
@@ -73,7 +50,7 @@ func getAddresses(cs *kubernetes.Clientset, endp *Endpoint) error {
 }
 
 func main() {
-	var cfg Config
+	var cfg config.Config
 
 	if err := cleanenv.ReadConfig(configPath, &cfg); err != nil {
 		log.Println(err)
